Refactor PBT and extract the runner
